Struct geom::angle::Angle

pub struct Angle(f64);

An angle, stored in radians.

Implementations

impl Angle

pub const ZERO: Angle

pub fn new_rads(rads: f64) -> Angle

Create an angle in radians.

pub fn degrees(degs: f64) -> Angle

Create an angle in degrees.

pub fn opposite(self) -> Angle

Invert the direction of this angle.

pub(crate) fn invert_y(self) -> Angle

pub fn rotate_degs(self, degrees: f64) -> Angle

Rotates this angle by some degrees.

pub fn normalized_radians(self) -> f64

Returns [0, 2pi)

pub fn normalized_degrees(self) -> f64

Returns [0, 360)

pub fn simple_shortest_rotation_towards(self, other: Angle) -> f64

Returns [-180, 180]

pub fn shortest_rotation_towards(self, other: Angle) -> Angle

Logically this returns [-180, 180], but keep in mind when we print this angle, it’ll normalize to be [0, 360].

pub fn approx_eq(self, other: Angle, within_degrees: f64) -> bool

True if this angle is within some degrees of another, accounting for rotation

pub fn approx_parallel(self, other: Angle, within_degrees: f64) -> bool

True if this angle is within some degrees of another, accounting for rotation and allowing two angles that point in opposite directions

pub fn reorient(self) -> Angle

I don’t know how to describe what this does. Use for rotating labels in map-space and making sure the text is never upside-down.

pub fn average(input: Vec<Angle>) -> Angle

Calculates the average of some angles.