Struct geom::find_closest::FindClosest

pub struct FindClosest<K> {
    geometries: BTreeMap<K, LineString<f64>>,
    quadtree: QuadTree<K>,
}

A quad-tree to quickly find the closest points to some polylines.

Fields
geometries: BTreeMap<K, LineString<f64>>quadtree: QuadTree<K>
Implementations

impl<K> FindClosest<K> where
K: Clone + Ord + Debug,

pub fn new(bounds: &Bounds) -> FindClosest<K>

Creates the quad-tree, limited to points contained in the boundary.

pub fn add(&mut self, key: K, pts: &[Pt2D])

Add an object to the quadtree, remembering some key associated with the points.

pub fn all_close_pts(
&self,
query_pt: Pt2D,
max_dist_away: Distance
) -> Vec<(K, Pt2D, Distance)>

For every object within some distance of a query point, return the (object’s key, point on the object’s polyline, distance away).

pub fn closest_pt(
&self,
query_pt: Pt2D,
max_dist_away: Distance
) -> Option<(K, Pt2D)>

Finds the closest point on the existing geometry to the query pt.