Trait geom::point::Position2D

pub trait Position2D<T> {
    fn x(&self) -> T;
    fn y(&self) -> T;
    fn x_mut(&mut self) -> &mut T;
    fn y_mut(&mut self) -> &mut T;
}

A type with x and y coordinates.

Required Methods

fn x(&self) -> T

Returns a copy of the x coordinate of the position.

fn y(&self) -> T

Returns a copy of the y coordinate of the position.

fn x_mut(&mut self) -> &mut T

Returns a mutable reference to the x coordinate of the position.

fn y_mut(&mut self) -> &mut T

Returns a mutable reference to the y coordinate of the position.