Struct geom::Bounds

pub struct Bounds {
    pub min_x: f64,
    pub min_y: f64,
    pub max_x: f64,
    pub max_y: f64,
}

Represents a rectangular boundary of Pt2D points.

Fields
min_x: f64min_y: f64max_x: f64max_y: f64
Implementations

impl Bounds

pub fn new() -> Bounds

A boundary including no points.

pub fn zero() -> Self

pub fn from(pts: &[Pt2D]) -> Bounds

Create a boundary covering some points.

pub fn from_polygons(polygons: &[Polygon]) -> Bounds

Create a boundary covering some polygons.

pub fn update(&mut self, pt: Pt2D)

Update the boundary to include this point.

pub fn union(&mut self, other: Bounds)

Unions two boundaries.

pub fn add_buffer(&mut self, sides: Distance)

Expand the existing boundary by some distance evenly on all sides.

pub fn scale(self, factor: f64) -> Self

Transform the boundary by scaling its corners.

pub fn contains(&self, pt: Pt2D) -> bool

True if the point is within the boundary.

pub fn as_bbox(&self) -> Rect

Converts the boundary to the format used by aabb_quadtree.

pub fn get_rectangle(&self) -> Polygon

Creates a rectangle covering this boundary.

pub fn to_circle(&self) -> Circle

Creates a circle centered in the middle of this boundary. Always uses the half width as a radius, so if the width and height don’t match, this is pretty meaningless.

pub fn width(&self) -> f64

The width of this boundary.

pub fn height(&self) -> f64

The height of this boundary.

pub fn center(&self) -> Pt2D

The center point of this boundary.