Struct geom::Line

pub struct Line(Pt2D, Pt2D);

A line segment.

Tuple Fields
0: Pt2D1: Pt2D
Implementations

impl Line

pub fn new(pt1: Pt2D, pt2: Pt2D) -> Result<Line>

Creates a line segment between two points, which must not be the same

pub fn must_new(pt1: Pt2D, pt2: Pt2D) -> Line

Equivalent to Line::new(pt1, pt2).unwrap(). Use this to effectively document an assertion at the call-site.

pub fn infinite(&self) -> InfiniteLine

Returns an infinite line passing through this line’s two points.

pub fn pt1(&self) -> Pt2D

Returns the first point in this line segment.

pub fn pt2(&self) -> Pt2D

Returns the second point in this line segment.

pub fn points(&self) -> Vec<Pt2D>

Returns the two points in this line segment.

pub fn to_polyline(&self) -> PolyLine

Returns a polyline containing these two points.

pub fn make_polygons(&self, thickness: Distance) -> Polygon

Returns a thick line segment.

pub fn length(&self) -> Distance

Length of the line segment

pub fn intersection(&self, other: &Line) -> Option<Pt2D>

If two line segments intersect – including endpoints – return the point where they hit. Undefined if the two lines have more than one intersection point!

pub fn crosses(&self, other: &Line) -> bool

Determine if two line segments intersect, but more so than just two endpoints touching.

pub fn intersection_infinite(&self, other: &InfiniteLine) -> Option<Pt2D>

If the line segment intersects with an infinite line – including endpoints – return the point where they hit. Undefined if the segment and infinite line intersect at more than one point!

pub fn shift_right(&self, width: Distance) -> Line

Perpendicularly shifts the line over to the right. Width must be non-negative.

pub fn shift_left(&self, width: Distance) -> Line

Perpendicularly shifts the line over to the left. Width must be non-negative.

pub fn shift_either_direction(&self, width: Distance) -> Line

Perpendicularly shifts the line to the right if positive or left if negative.

pub fn reversed(&self) -> Line

Returns a reversed line segment

pub fn angle(&self) -> Angle

The angle of the line segment, from the first to the second point

pub fn dist_along(&self, dist: Distance) -> Result<Pt2D>

Returns a point along the line segment, unless the distance exceeds the segment’s length.

pub fn must_dist_along(&self, dist: Distance) -> Pt2D

Equivalent to self.dist_along(dist).unwrap(). Use this to document an assertion at the call-site.

pub fn unbounded_dist_along(&self, dist: Distance) -> Pt2D

pub fn unbounded_percent_along(&self, percent: f64) -> Pt2D

pub fn percent_along(&self, percent: f64) -> Result<Pt2D>

pub fn slice(&self, from: Distance, to: Distance) -> Result<Line>

pub fn percent_slice(&self, from: f64, to: f64) -> Result<Line>

Returns a subset of this line, with two percentages along the line’s length.

pub fn middle(&self) -> Result<Pt2D>

pub fn contains_pt(&self, pt: Pt2D) -> bool

pub fn dist_along_of_point(&self, pt: Pt2D) -> Option<Distance>

pub fn percent_along_of_point(&self, pt: Pt2D) -> Option<f64>