Struct geom::LonLat

pub struct LonLat {
    longitude: NotNan<f64>,
    latitude: NotNan<f64>,
}

Represents a (longitude, latitude) point.

Fields
longitude: NotNan<f64>latitude: NotNan<f64>
Implementations

impl LonLat

pub fn new(lon: f64, lat: f64) -> LonLat

Note the order of arguments!

pub fn x(self) -> f64

Returns the longitude of this point.

pub fn y(self) -> f64

Returns the latitude of this point.

pub fn to_pt(self, b: &GPSBounds) -> Pt2D

Transform this to a world-space point. Can go out of bounds.

pub(crate) fn gps_dist(self, other: LonLat) -> Distance

Returns the Haversine distance to another point.

pub fn fast_dist(self, other: LonLat) -> NotNan<f64>

Pretty meaningless units, for comparing distances very roughly

pub fn read_osmosis_polygon(path: &str) -> Result<Vec<LonLat>>

Parses a file in the https://wiki.openstreetmap.org/wiki/Osmosis/Polygon_Filter_File_Format and returns all points.

pub fn write_osmosis_polygon(path: &str, pts: &[LonLat]) -> Result<()>

Writes a set of points to a file in the https://wiki.openstreetmap.org/wiki/Osmosis/Polygon_Filter_File_Format. The input should be a closed ring, with the first and last point matching.

pub fn center(pts: &[LonLat]) -> LonLat

Finds the average of a set of coordinates.

pub fn parse_wkt_linestring(raw: &str) -> Option<Vec<LonLat>>

Parses a WKT-style line-string into a list of coordinates.

pub fn parse_geojson_polygons(
raw: String
) -> Result<Vec<(Vec<LonLat>, Option<String>)>>

Extract polygons from a raw GeoJSON string. For multipolygons, only returns the first member. If the GeoJSON feature has a property called name, this will also be returned.

pub fn to_geojson(self) -> Geometry