Struct geom::PolyLine

pub struct PolyLine {
    pts: Vec<Pt2D>,
    length: Distance,
}

Fields
pts: Vec<Pt2D>length: Distance
Implementations

impl PolyLine

pub fn new(pts: Vec<Pt2D>) -> Result<PolyLine, String>

pub fn must_new(pts: Vec<Pt2D>) -> PolyLine

pub fn unchecked_new(pts: Vec<Pt2D>) -> PolyLine

Doesn't check for duplicates. Use at your own risk.

pub fn deduping_new(pts: Vec<Pt2D>) -> Result<PolyLine, String>

First dedupes adjacent points

pub fn to_thick_boundary(
&self,
self_width: Distance,
boundary_width: Distance
) -> Option<Polygon>

pub fn reversed(&self) -> PolyLine

pub fn extend(self, other: PolyLine) -> Result<PolyLine, String>

pub fn must_extend(self, other: PolyLine) -> PolyLine

pub fn must_push(self, pt: Pt2D) -> PolyLine

pub fn append(first: Vec<Pt2D>, second: Vec<Pt2D>) -> Result<Vec<Pt2D>, String>

One or both args might be empty.

pub fn points(&self) -> &Vec<Pt2D>

pub fn into_points(self) -> Vec<Pt2D>

pub fn lines(&self) -> Vec<Line>

pub fn length(&self) -> Distance

pub fn slice(
&self,
start: Distance,
end: Distance
) -> Result<(PolyLine, Distance), String>

Returns the excess distance left over from the end

pub fn exact_slice(&self, start: Distance, end: Distance) -> PolyLine

No excess leftover distance allowed.

pub fn maybe_exact_slice(
&self,
start: Distance,
end: Distance
) -> Result<PolyLine, String>

pub fn first_half(&self) -> PolyLine

pub fn second_half(&self) -> PolyLine

pub fn dist_along(&self, dist_along: Distance) -> Result<(Pt2D, Angle), String>

pub fn must_dist_along(&self, dist_along: Distance) -> (Pt2D, Angle)

pub fn middle(&self) -> Pt2D

pub fn first_pt(&self) -> Pt2D

pub fn last_pt(&self) -> Pt2D

pub fn first_line(&self) -> Line

pub fn last_line(&self) -> Line

pub fn shift_right(&self, width: Distance) -> Result<PolyLine, String>

pub fn must_shift_right(&self, width: Distance) -> PolyLine

pub fn shift_left(&self, width: Distance) -> Result<PolyLine, String>

pub fn must_shift_left(&self, width: Distance) -> PolyLine

fn shift_with_corrections(&self, width: Distance) -> Result<PolyLine, String>

fn shift_with_sharp_angles(
&self,
width: Distance,
miter_threshold: f64
) -> Vec<Pt2D>

pub fn make_polygons(&self, width: Distance) -> Polygon

pub fn make_polygons_with_miter_threshold(
&self,
width: Distance,
miter_threshold: f64
) -> Polygon

pub fn exact_dashed_polygons(
&self,
width: Distance,
dash_len: Distance,
dash_separation: Distance
) -> Vec<Polygon>

pub fn dashed_lines(
&self,
width: Distance,
dash_len: Distance,
dash_separation: Distance
) -> Vec<Polygon>

Don't draw the dashes too close to the ends.

pub fn make_arrow(&self, thickness: Distance, cap: ArrowCap) -> Polygon

pub fn make_double_arrow(&self, thickness: Distance, cap: ArrowCap) -> Polygon

pub fn dashed_arrow(
&self,
width: Distance,
dash_len: Distance,
dash_separation: Distance,
cap: ArrowCap
) -> Vec<Polygon>

pub fn intersection(&self, other: &PolyLine) -> Option<(Pt2D, Angle)>

Also return the angle of the line where the hit was found

pub fn intersection_infinite(&self, other: &InfiniteLine) -> Option<Pt2D>

pub fn get_slice_ending_at(&self, pt: Pt2D) -> Option<PolyLine>

Panics if the pt is not on the polyline. Returns None if the point is the first point (meaning the slice is empty).

pub fn get_slice_starting_at(&self, pt: Pt2D) -> Option<PolyLine>

Returns None if the point is the last point.

pub fn dist_along_of_point(&self, pt: Pt2D) -> Option<(Distance, Angle)>

pub fn trim_to_endpts(&self, pt1: Pt2D, pt2: Pt2D) -> PolyLine

pub fn get_bounds(&self) -> Bounds

pub fn extend_to_length(&self, min_len: Distance) -> PolyLine

If the current line is at least this long, return it. Otherwise, extend the end of it, following the angle of the last line.