Struct geom::tessellation::Tessellation

pub struct Tessellation {
    pub(crate) points: Vec<Pt2D>,
    indices: Vec<u16>,
}

A tessellated polygon, ready for rendering.

Fields
points: Vec<Pt2D>

These points aren’t in any meaningful order. It’s not generally possible to reconstruct a Polygon from this.

indices: Vec<u16>

Groups of three indices make up the triangles

Implementations

impl Tessellation

pub fn new(points: Vec<Pt2D>, indices: Vec<usize>) -> Self

pub fn from_ring(points: Vec<Pt2D>) -> Self

The points are not necessarily a Ring, which has strict requirements about no duplicate points. We can render various types of invalid polygon.

pub fn consume(self) -> (Vec<Pt2D>, Vec<u16>)

Returns (points, indices) for rendering

pub fn triangles(&self) -> Vec<Triangle>

pub fn get_bounds(&self) -> Bounds

pub fn center(&self) -> Pt2D

pub(crate) fn transform<F: Fn(&Pt2D) -> Pt2D>(&mut self, f: F)

pub fn translate(&mut self, dx: f64, dy: f64)

pub fn scale(&mut self, factor: f64)

pub fn scale_xy(&mut self, x_factor: f64, y_factor: f64)

pub fn rotate(&mut self, angle: Angle)

pub fn rotate_around(&mut self, angle: Angle, pivot: Pt2D)

pub fn inplace_multi_transform( &mut self, scale: f64, translate_x: f64, translate_y: f64, rotate: Angle, pivot: Pt2D )

Equivalent to self.scale(scale).translate(translate_x, translate_y).rotate_around(rotate, pivot), but modifies the polygon in-place and is faster.

pub fn union(self, other: Self) -> Self

pub fn union_all(list: Vec<Self>) -> Self

pub fn to_geojson(&self, gps: Option<&GPSBounds>) -> Geometry

Produces a GeoJSON multipolygon consisting of individual triangles. Optionally map the world-space points back to GPS.

fn to_geo(&self) -> Polygon

pub fn difference(&self, other: &Tessellation) -> Result<Vec<Polygon>>