Struct geom::Duration

pub struct Duration(f64);

A duration, in seconds. Can be negative.

Implementations

impl Duration

pub const ZERO: Duration

pub const EPSILON: Duration

pub fn seconds(value: f64) -> Duration

Creates a duration in seconds.

pub fn minutes(mins: usize) -> Duration

Creates a duration in minutes.

pub fn hours(hours: usize) -> Duration

Creates a duration in hours.

pub fn f64_minutes(mins: f64) -> Duration

Creates a duration in minutes.

pub const fn const_seconds(value: f64) -> Duration

pub(crate) fn to_u64(self) -> u64

pub(crate) fn from_u64(x: u64) -> Duration

pub fn abs(&self) -> Self

pub fn inner_seconds(self) -> f64

Returns the duration in seconds. Prefer working in typesafe Durations.

fn get_parts(self) -> (usize, usize, usize, usize)

Splits the duration into (hours, minutes, seconds, deciseconds).

pub fn parse(string: &str) -> Result<Duration>

Parses a duration such as “3:00” to Duration::minutes(3).

pub fn epsilon_eq(self, other: Duration) -> bool

If two durations are within this amount, they’ll print as if they’re the same.

pub fn realtime_elapsed(since: Instant) -> Duration

Returns the duration elapsed from this moment in real time.

pub fn round_up(self, multiple: Duration) -> Duration

Rounds a duration up to the nearest whole number multiple.

pub fn num_minutes_rounded_up(self) -> usize

Returns the duration as a number of minutes, rounded up.

pub fn make_intervals_for_max(
self,
num_labels: usize
) -> (Duration, Vec<Duration>)

Returns (rounded max, the boundaries)

pub fn to_rounded_string(self, precision: usize) -> String

Shows only the largest unit (hours, minute, seconds), rounded to precision decimal points.

use geom::Duration;
assert_eq!(Duration::seconds(3600.0).to_rounded_string(0), "1hr");
assert_eq!(Duration::seconds(3600.0).to_rounded_string(1), "1.0hr");
assert_eq!(Duration::seconds(7800.0).to_rounded_string(0), "2hr");
assert_eq!(Duration::seconds(800.0).to_rounded_string(1), "13.3min");
assert_eq!(Duration::seconds(-800.0).to_rounded_string(1), "-13.3min");
assert_eq!(Duration::seconds(0.0).to_rounded_string(0), "0");
assert_eq!(Duration::seconds(12.5).to_rounded_string(1), "12.5s");
assert_eq!(Duration::seconds(12.5).to_rounded_string(2), "12.50s");

pub fn to_string(self, fmt: &UnitFmt) -> String

Describes the duration according to formatting rules.